Added classes and started with the basics in every class.

+++ src/main/java/com/example/watchtime/api/TVShowController.java
@@ -17,7 +17,7 @@ public class TVShowController {
     }
 
     @GetMapping("/tvshow/{id}")
-    public TVShow getTVShowById(@PathVariable("id") Integer id) {
+    public TVShow getTVShowById(@PathVariable("id") Long id) {
         return tvShowService.getTVShowById(id);
     }
 }
