+++ src/main/java/com/example/watchtime/repository/TVShowRepository.java
@@ -3,5 +3,5 @@ package com.example.watchtime.repository;
 import com.example.watchtime.model.TVShow;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TVShowRepository extends CrudRepository<TVShow, Integer> {
+public interface TVShowRepository extends CrudRepository<TVShow, Long> {
 }
