+++ src/main/java/com/example/watchtime/repository/WatchlistRepository.java
@@ -3,5 +3,5 @@ package com.example.watchtime.repository;
 import com.example.watchtime.model.Watchlist;
 import org.springframework.data.repository.CrudRepository;
 
-public interface WatchlistRepository extends CrudRepository<Watchlist, Integer> {
+public interface WatchlistRepository extends CrudRepository<Watchlist, Long> {
 }
